make fire warg a .h file
